Prototype/backend/requirements.txt
# This file contains any Python dependences we'll need for the project
# Add any additional packages you need here
# Install by running
# pip install -r requirements.txt

# Werkzeug has to be version 2.12 or less to fix a bug with flask-restx 
Werkzeug <= 2.1.2   
flask == 2.1.2   
flask-restx >= 0.5.1
Flask-Cors==3.0.10
python-dotenv==0.21.0
requests==2.28.1
pandas==1.5.2
